# openapi/v0.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '0.1.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Because the Twext compiler does not embed manifest metadata into its
    compiled output (unlike the old Warp/Fluorite `Warp.meta` convention),
    every publish call must submit its manifest fields explicitly as JSON
    alongside the compiled code — TwextHub never parses the JS blob to
    recover metadata.

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they call
    `POST /terms/accept`.

    Bearer tokens come in two kinds. **Sessions** are created by
    `/auth/login` or `/auth/signup`, expire after 7 days, and implicitly
    satisfy every scope. **Automation tokens** are created explicitly via
    `POST /tokens`, are scoped to just `publish` and/or `yank` for CI use
    (e.g. a `twext publish` GitHub Action), and never expire unless the
    caller sets one. Only a session may manage sessions or automation
    tokens, act on `/admin/*` endpoints, or change account settings —
    an automation token calling any of those gets a `403`.
servers:
...
    variables:
...
security:
  - BearerAuth: []
paths:
...
  /auth/signup:
    post:
      summary: Create an account
      operationId: signup
      tags: [Open]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace: { $ref: '#/components/schemas/Namespace' }
                displayName: { type: string, maxLength: 80 }
                password: { type: string, minLength: 8, writeOnly: true }
      responses:
        '201':
          description: Account created. The first account ever created becomes an admin.
          content:
            application/json:
              schema:
                type: object
                required: [user, token]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/login:
    post:
      summary: Log in
      operationId: login
      tags: [Open]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace: { $ref: '#/components/schemas/Namespace' }
                password: { type: string, writeOnly: true }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                required: [user, token]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/logout:
    post:
      summary: Revoke the current bearer token
      operationId: logout
      tags: [Userspace]
      description: Equivalent to `DELETE /sessions/{id}` for the session that authenticated this request.
      responses:
        '204': { description: Token revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      summary: Get the authenticated account
      operationId: getMe
      tags: [Userspace]
      responses:
        '200':
          description: The caller's own account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------------------------------------------- Sessions & automation tokens
...
  /extensions:
    get:
      summary: List published extensions
      operationId: listExtensions
      tags: [Open]
      security: []
      description: Latest published version per `(namespace, id)`, newest first.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated extensions
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ExtensionSummary' }
                  pagination: { $ref: '#/components/schemas/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /search:
    get:
      summary: Search published extensions
      operationId: searchExtensions
      tags: [Open]
      security: []
      description: Matches id, name, namespace, and description. Omit `query` for a recency-ordered fallback.
      parameters:
        - name: query
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated search results
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ExtensionSummary' }
                  pagination: { $ref: '#/components/schemas/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /terms/accept:
    post:
      summary: Accept the current Terms of Service
      operationId: acceptTerms
      tags: [Userspace]
      description: Records acceptance of whichever `version` of the Terms of Service is current at the time of the call.
      responses:
        '204': { description: Recorded }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /admin/terms:
    patch:
      summary: Publish a new Terms of Service version
      operationId: updateTerms
      tags: [Adminspace]
      description: Replaces the current Terms of Service body and bumps `version`. Previously-accepted versions no longer satisfy the acceptance gate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, description: Markdown source. }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LegalDocument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /admin/privacy:
    patch:
      summary: Publish a new Privacy Policy version
      operationId: updatePrivacyPolicy
      tags: [Adminspace]
      description: Replaces the current Privacy Policy body and bumps `version`. Does not itself require re-acceptance (only Terms updates gate `Userspace`/`Adminspace` calls).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, description: Markdown source. }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LegalDocument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /versions:
    get:
      summary: Cross-owner moderation queue
      operationId: listVersionsForReview
      tags: [Adminspace]
      parameters:
        - name: status
          in: query
          required: true
          schema: { type: string, enum: [pending] }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated pending versions, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PendingVersion' }
                  pagination: { $ref: '#/components/schemas/CursorPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque pagination cursor from a previous response.
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 50 }
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
    ExtensionId:
      type: string
      pattern: '^[a-z0-9]{1,64}$'
      description: >
        Matches TurboWarp's own documented rule for extension `id`s: only
        `a-z` and `0-9`, no spaces or special characters. The 64-char cap
        is TwextHub's own storage/URL-length limit, not a TurboWarp
        requirement. twext.yml's `extension.id` field has no pattern
        constraint today; the `twext` CLI's own validation should be
        tightened to match this rather than the other way around.
    Role:
      type: string
      enum: [admin, normal]
...
    User:
      type: object
      required: [namespace, displayName, role, hasPublished, createdAt, termsAcceptedVersion]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        hasPublished: { type: boolean }
        createdAt: { type: string, format: date-time, readOnly: true }
        termsAcceptedVersion:
          type: ['integer', 'null']
          description: Terms of Service `version` this account last accepted, or `null` if never accepted.
    LegalDocument:
      type: object
      required: [version, body, updatedAt]
      properties:
        version: { type: integer, description: Increments on every publish of a new version. }
        body: { type: string, description: Markdown source. }
        updatedAt: { type: string, format: date-time }
...
    ExtensionSummary:
      type: object
      required: [namespace, id, name, version, description, publishedAt]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        id: { $ref: '#/components/schemas/ExtensionId' }
        name: { type: string }
        version: { type: string }
        description: { type: string }
        publishedAt: { type: string, format: date-time }
...
    VersionInfo:
      type: object
      required: [namespace, id, version, status, createdAt]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        id: { $ref: '#/components/schemas/ExtensionId' }
        version: { type: string }
        status: { type: string, enum: [pending, published, rejected, yanked] }
        name: { type: string }
        license: { type: string }
        description: { type: string }
        author: { type: string }
        createdAt: { type: string, format: date-time }
        publishedAt:
          { type: string, format: date-time, description: Absent while pending or rejected. }
        dist:
          type: object
          description: Absent while pending or rejected.
          properties:
            downloadUrl: { type: string, format: uri }
    PendingVersion:
      allOf:
        - $ref: '#/components/schemas/VersionInfo'
        - type: object
          required: [ownerNamespace]
          properties:
            ownerNamespace: { $ref: '#/components/schemas/Namespace' }
...
    CursorPage:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: ['string', 'null'] }
        hasMore: { type: boolean }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          { type: string, format: uri, example: 'https://hub.twext.dev/errors/validation-error' }
        title: { type: string, example: 'Validation Error' }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: >
        Authenticated, but not allowed to perform this action — e.g. not the
        resource owner or an admin, a namespace mismatch on publish, the
        current Terms of Service haven't been accepted yet
        (see `POST /terms/accept`), an automation token missing the
        required scope, or an automation token used against a
        session-only endpoint (session/token management, `/admin/*`,
        account settings).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    Conflict:
      description: Resource already exists
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    UnprocessableEntity:
      description: Semantically invalid input (e.g. non-SemVer version, id mismatch)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer